+++ tests/test_ctx.py
@@ -13,7 +13,7 @@ from rxcat import (
     Transport,
     ok,
 )
-from rxcat.code_ext import get_registered_codeid_by_type
+from rxcat.code_ext import get_regd_codeid_by_type
 from tests.conftest import (
     Mock_1,
     MockConn,
@@ -33,11 +33,11 @@ async def test_subfn(server_bus: ServerBus):
     await asyncio.wait_for(conn.client__recv(), 1)
     await conn.client__send({
         "sid": uuid4(),
-        "datacodeid": (await get_registered_codeid_by_type(Mock_1)).eject(),
+        "datacodeid": (await get_regd_codeid_by_type(Mock_1)).eject(),
         "num": 1
     })
     rmsg = await asyncio.wait_for(conn.client__recv(), 1)
-    assert rmsg["datacodeid"] == get_registered_codeid_by_type(ok)
+    assert rmsg["datacodeid"] == get_regd_codeid_by_type(ok)
     conn_task.cancel()
 
 async def test_rpc(server_bus: ServerBus):
@@ -49,18 +49,18 @@ async def test_rpc(server_bus: ServerBus):
     conn_task = asyncio.create_task(server_bus.conn(conn))
     # recv welcome
     await asyncio.wait_for(conn.client__recv(), 1)
-    ServerBus.register_rpc(srpc__update_email).eject()
+    ServerBus.reg_rpc(srpc__update_email).eject()
     rpc_token = uuid4()
     rpc_key = "srpc__update_email:" + rpc_token
     await conn.client__send({
         "sid": uuid4(),
-        "datacodeid": (await get_registered_codeid_by_type(SrpcSend)).eject(),
+        "datacodeid": (await get_regd_codeid_by_type(SrpcSend)).eject(),
         "key": rpc_key,
         "args": {"username": "test_username", "email": "test_email"}
     })
     rmsg = await asyncio.wait_for(conn.client__recv(), 1)
     assert rmsg["datacodeid"] == \
-        (await get_registered_codeid_by_type(SrpcRecv)).eject()
+        (await get_regd_codeid_by_type(SrpcRecv)).eject()
 
     conn_task.cancel()
 
@@ -81,7 +81,7 @@ async def test_rpc_custom_ctx_manager():
             Transport(
                 is_server=True,
                 conn_type=MockConn,
-                server__register_process="none")
+                server__reg_process="none")
         ],
         sub_ctxfn=get_mock_ctx_manager_for_msg))
 
@@ -93,17 +93,17 @@ async def test_rpc_custom_ctx_manager():
     conn_task = asyncio.create_task(server_bus.conn(conn))
     # recv welcome
     await asyncio.wait_for(conn.client__recv(), 1)
-    ServerBus.register_rpc(srpc__update_email).eject()
+    ServerBus.reg_rpc(srpc__update_email).eject()
     rpc_token = uuid4()
     rpc_key = "srpc__update_email:" + rpc_token
     await conn.client__send({
         "sid": uuid4(),
-        "datacodeid": (await get_registered_codeid_by_type(SrpcSend)).eject(),
+        "datacodeid": (await get_regd_codeid_by_type(SrpcSend)).eject(),
         "key": rpc_key,
         "args": {}
     })
     rmsg = await asyncio.wait_for(conn.client__recv(), 1)
     assert rmsg["datacodeid"] == \
-        (await get_registered_codeid_by_type(SrpcRecv)).eject()
+        (await get_regd_codeid_by_type(SrpcRecv)).eject()
 
     conn_task.cancel()